requirements.txt
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=0.19.0
chromadb>=0.3.0
tiktoken>=0.3.0
numpy>=1.24.0
pandas>=2.0.0
streamlit>=1.24.0